Move GymForm to vue
