+++ frontend/src/index.js
@@ -2,6 +2,7 @@ import HelloWorld from "@/components/HelloWorld";
 import CounterModule from "@/vuex/vuex_module_counter";
 import {createAppInEl, createSharedStore} from "@/util/create_app_utils";
 import Modal from "@/components/Modal";
+import GymForm from "@/components/GymForm";
 
 // Initialize our shared store including the counter module
 let store = createSharedStore({
@@ -11,4 +12,4 @@ let store = createSharedStore({
 // Create and mount our apps
 createAppInEl(HelloWorld, store, "#hello_world_a", []);
 createAppInEl(HelloWorld, store, "#hello_world_b", []);
-createAppInEl({}, null, "#app", [Modal]);
+createAppInEl({}, null, "#app", [GymForm, Modal]);
