+++ python_anywhere/bouldern/templates/bouldern/gym_form.html
@@ -11,27 +11,22 @@
     <div id="hello_world_b" data-msg="{{ msg1 }}" style="width: 49%; display: inline-block"></div>
 {% endwith %}
 
-<form action="/bouldern/add-gym/" method="post" enctype="multipart/form-data">
-    {% csrf_token %}
-    {{ gym_form }}
-    <h3>Difficulty Levels</h3>
-    {{ difficulty_level_formset.management_form }}
-    <div id="difficulty-levels">
-        {% for difficulty_level_form in difficulty_level_formset %}
-            <div id="difficulty-level">
-                {{ difficulty_level_form }}
-                {{ difficulty_level_form.media }}
-            </div>
-        {% endfor %}
-    </div>
-    <button type="button" id="add-level-button">Add Level</button>
-    <input type="submit" value="Submit">
-</form>
 <div id="app">
-    <Modal>
-        {% include 'bouldern/color_form.html' with form=color_form %}
-    </Modal>
+    <gym-form>
+        <template #gym-form>{% csrf_token %}{{ gym_form }}</template>
+        <template #difficulty-levels>
+            {{ difficulty_level_formset.management_form }}
+            {% for difficulty_level_form in difficulty_level_formset %}
+                <div id="difficulty-level">
+                    {{ difficulty_level_form }}
+                </div>
+            {% endfor %}
+        </template>
+    </gym-form>
+    {#Todo: change tag names all to lower case#}
+    <Modal>{% include 'bouldern/color_form.html' with form=color_form %}</Modal>
 </div>
+{{ difficulty_level_formset.0.media }}
 {{ color_form.media }}
 {% render_bundle 'index' %}
 </body>
